LL-2496 abstract ui from tron delegation
